feat: new settings page
